Fully detach webscraper
