+++ scripts/webscraper.py
@@ -1,135 +1,90 @@
+# WARNING: Running this script will overwrite the class_data table!
+
 from bs4 import BeautifulSoup
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import Select
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions
 import pandas as pd
-
-# for multithreading
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_cors import CORS
+import pandas as pd
+import sys
+import os
 import concurrent.futures
+
+# add service directory to path
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'service')))
+from database import db, ClassData
+
+# IMPORTANT: REPLACE THIS WITH POSTGRESQL URL if you want to edit the actual database
+service_directory = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'service'))
+db_path = os.path.join(service_directory, 'table.db')
+DATABASE_URL = f'sqlite:///{db_path}'
 
 NAU_CLASS_DISTRIBUTION_URL = "https://www7.nau.edu/pair/reports/ClassDistribution"
 
-# get data iteratively
-def get_all_grade_distribution_data():
-    # store sata here
-    class_distr_data = list()
+# fill database with class data
+def main():
+    # set up database
+    app = Flask(__name__)
+    CORS(app)
 
-    service = Service()
-    options = webdriver.ChromeOptions()
-    options.add_argument("--disable-gpu")
-    options.add_argument("--disable-extensions")
-    options.add_argument("--disable-infobars")
-    options.add_argument("--start-maximized")
-    options.add_argument("--disable-notifications")
-    options.add_argument('--headless')
-    options.add_argument('--no-sandbox')
-    options.add_argument('--disable-dev-shm-usage')
-    driver = webdriver.Chrome(service=service, options=options)
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
-    driver.get(NAU_CLASS_DISTRIBUTION_URL)
+    app.config['SQLALCHEMY_DATABASE_URI'] = DATABASE_URL
 
-    # wait until subject term is done loading
-    WebDriverWait(driver, 10).until(expected_conditions.presence_of_element_located((By.NAME, "ctl00$MainContent$TermList")))
-    # select term by name (skip default)
-    term_select_element = Select(driver.find_element(By.NAME, "ctl00$MainContent$TermList"))
-    for term_idx in range(1, len(term_select_element.options)):
-        # get the selected term name
-        term = term_select_element.options[term_idx].text
+    # bind the database to backend
+    db.init_app(app)
 
-        # select term
-        term_select_element.select_by_index(term_idx)
+    # create database tables
+    with app.app_context():
+        fetch_grade_distribution_data(db)
 
-        # wait until subject select is done loading
-        WebDriverWait(driver, 10).until(expected_conditions.presence_of_element_located((By.NAME, "ctl00$MainContent$SubjectList")))
-         # select subject by name
-        subject_select_element = Select(driver.find_element(By.NAME, "ctl00$MainContent$SubjectList"))
+# fill database with class data
+def fetch_grade_distribution_data(db: SQLAlchemy):
+    grade_distribution_df = get_all_grade_distribution_data_parallel()
 
-        """""
-        # ---------------------------
-        # GRAB ALL CLASSES
-        # ---------------------------
-        for subject_idx in range(0, len(subject_select_element.options)):
-            # get the selected subject name
-            subject = subject_select_element.options[subject_idx].text
-
-            # select subject
-            subject_select_element.select_by_index(subject_idx)
-        # ---------------------------
-        """
-        # ---------------------------
-        # ONLY GRABBING CS CLASSES
-        # ---------------------------
-        subject_idx = 0
-        subject = "CS"
-        if subject in [option.text for option in subject_select_element.options]:
-            subject_select_element.select_by_visible_text(subject)
-        # ---------------------------
-
-            # click submit
-            driver.find_element(By.ID, "MainContent_Button1").click()
-
-            # wait until the first table is done loading (max 10 seconds)
-            WebDriverWait(driver, 10).until(expected_conditions.presence_of_element_located((By.TAG_NAME, "table")))
-
-            # get the html from curent page
-            html = driver.page_source
-
-            # use beautful soup to parse
-            soup = BeautifulSoup(html, 'html.parser')
-
-            tables = soup.find_all("table")
-
-            # loop through each table
-            for table_tag in tables:
-
-                # extract the table
-                table = table_tag.find("tbody")
-
-                if table:
-                    
-                    # loop through each row in the table
-                    for row_idx, tr in enumerate(table.find_all("tr")):
-                        row_data = [td.text for td in tr.find_all("td")]
-
-                        # skip any rows that are blank
-                        if row_data:
-                            # assume first row is the header
-                            if row_idx == 0:
-                                # only need to get the header once
-                                if term_idx == 1 and subject_idx == 0:
-                                    col_headers = ["Semester", "Subject"] + row_data
-                            # otherwise add row to table
-                            else:
-                                # concat semester and subject info to row data
-                                class_distr_data.append([term, subject] + row_data)
-
-                else:
-                    print("Table not found.")
-
-            """""
-            # go back to select next subject
-            driver.back()
-            # refetch dropdown
-            WebDriverWait(driver, 10).until(expected_conditions.presence_of_element_located((By.NAME, "ctl00$MainContent$SubjectList")))
-            subject_select_element = Select(driver.find_element(By.NAME, "ctl00$MainContent$SubjectList"))
-            """""
-
-        # go back to select next term
-        driver.back()
-        # refetch dropdown
-        WebDriverWait(driver, 10).until(expected_conditions.presence_of_element_located((By.NAME, "ctl00$MainContent$TermList")))
-        term_select_element = Select(driver.find_element(By.NAME, "ctl00$MainContent$TermList"))
+    data_to_add = []
     
-    # create data frame
-    class_distr_table = pd.DataFrame(class_distr_data, columns=col_headers)
+    # create database object for each row
+    # note: I want to use .tosql, but it won't work
+    for idx, row in grade_distribution_df.iterrows():
+        data = ClassData(
+            semester = row["Semester"],
+            subject = row["Subject"],
+            class_name = row["Class"],
+            section = row["Section"],
+            class_nbr = int(row["Class NBR "]),
+            instructor_name = row["Instructor Name"],
+            a = int(row["A"]),
+            b = int(row["B"]),
+            c = int(row["C"]),
+            d = int(row["D"]),
+            f = int(row["F"]),
+            au = int(row["AU"]),
+            p = int(row["P"]),
+            ng = int(row["NG"]),
+            w = int(row["W"]),
+            i = int(row["I"]),
+            ip = int(row["IP"]),
+            pending = int(row["Pending"]),
+            total = int(row["Total"])
+        )
+        data_to_add.append(data)
+    
+    # drop current class_data table
+    db.create_all()
+    with db.session.begin():
+        ClassData.__table__.drop(db.engine)
+    db.create_all()
 
-    driver.quit()
-
-    return class_distr_table
+    # add all data objects to database
+    db.session.bulk_save_objects(data_to_add)
+    db.session.commit()
 
 # get data in parallel
 def get_all_grade_distribution_data_parallel():
@@ -170,6 +125,7 @@ def get_all_grade_distribution_data_parallel():
     col_headers = ["Semester", "Subject", "Class", "Section", "Class NBR ", "Instructor Name", "A", "B", "C", "D", "F", "AU", "P", "NG", "W", "I", "IP", "Pending", "Total"]
     class_distr_table = pd.DataFrame(all_class_distr_data, columns=col_headers)
 
+    # split rows that have multiple instructor names so they each get their own row with the same data for the calss
     class_distr_table["Instructor Name"] = class_distr_table["Instructor Name"].str.split('/')
     class_distr_table = class_distr_table.explode("Instructor Name").reset_index()
 
@@ -180,9 +136,9 @@ def get_all_grade_distribution_data_parallel():
 
     
 def fetch_info_single_term(term):
+    print(f"Scraping data for {term}...")
     # store data here
     term_class_distr_data = list()
-    col_headers = list()
 
     # open driver for each thread
     service = Service()
@@ -280,4 +236,4 @@ def fetch_info_single_term(term):
     return term_class_distr_data
 
 if __name__ == "__main__":
-    print(get_all_grade_distribution_data_parallel())
+    main()
