+++ service/Dockerfile
@@ -1,38 +1,8 @@
 # service/Dockerfile
-FROM --platform=linux/amd64 python:3.9
-#FROM python:3.9
+FROM python:3.9
 
 # For MacOS
 # FROM --platform=linux/amd64  python:3.9
-
-# add google chrome and chrome driver to executables for webscraping
-# Adding trusting keys to apt for repositories
-RUN wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
-
-# Adding Google Chrome to the repositories
-RUN sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
-
-# Updating apt to see and install Google Chrome
-RUN apt-get -y update
-
-# Magic happens
-RUN apt-get install -y google-chrome-stable
-
-# Installing Unzip
-RUN apt-get install -yqq unzip
-
-# Download the Chrome Driver
-RUN wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
-RUN sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
-RUN apt-get -y update
-RUN apt-get install -y google-chrome-stable
-
-# install chromedriver to bin
-# version may need to be changes
-RUN wget https://edgedl.me.gvt1.com/edgedl/chrome/chrome-for-testing/130.0.6723.91/linux64/chromedriver-linux64.zip \
-  && unzip chromedriver-linux64.zip && rm -dfr chromedriver_linux64.zip \
-  && mv /chromedriver-linux64/chromedriver /usr/bin/chromedriver \
-  && chmod +x /usr/bin/chromedriver
 
 WORKDIR /app
