+++ service/app.py
@@ -1,8 +1,6 @@
 from flask import Flask, request, jsonify
 from flask_cors import CORS
-from database import db, fetch_grade_distribution_data
-import threading
-import os
+from database import db
 import logging
 import sys
 from search import search_for
@@ -14,13 +12,6 @@ from rating import get_average_rating, add_rating
 
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
 logger = logging.getLogger("service")
-
-# turn to true to start filling the database with class information when the server starts
-FILL_DB_WITH_CLASS_DATA = False
-
-# only fill if sqlite database does not already exists on start up
-if not os.path.isfile('table.db'):
-    FILL_DB_WITH_CLASS_DATA = True
 
 # ====================================
 # Create app
@@ -261,16 +252,5 @@ def post_rating_route():
 
 # ====================================
 
-def fill_db_with_class_data():
-    logger.info("\nWebscraper running...")
-    # only run this to fill database with class data
-    with app.app_context():
-        fetch_grade_distribution_data(db)
-    logger.info("\nWebscraper finished...")
-
 if __name__ == '__main__':
-    if FILL_DB_WITH_CLASS_DATA:
-        thread = threading.Thread(target=fill_db_with_class_data)
-        thread.start()
-
     app.run(host='0.0.0.0', port=8080)
